Split resolver import absence tests
